requirements.txt
matplotlib
pandas
lxml
openpyxl
ipykernelpyperclip